pytest/api-testing/pyproject.toml
[tool.poetry]
name = "api-testing"
version = "0.1.0"
description = ""
authors = ["Jon Whittlestone <dev@howapped.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
pytest-mock = "^3.12.0"
requests = "^2.31.0"
responses = "^0.24.1"
pytest-regressions = "^2.5.0"
vcrpy = "^5.1.0"
pytest-vcr = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
